[0.055]: [+:>>] MAJOR update. Many changes/corrections...
1) CCEXP contructor: Corrected order + SavingFile inititialize to 0 without
sprintf (avoid all warnings)

2) Converting -1 of size_t to std::numeric_limits<std::size_t>::max() when
comparing. (Avoid warnings / errors)

3) Use snprintf() instead of sprintf() (Avoid Memory Leaking)

4) API functions that return size_t on errors must return '0' (avoid overflow on
uncatched errors by accident).

5) getRow() and getVal() does not return anymore references, but pointers.
Thus in case of error, values are NULL.

6) Open/Close return Void

7) Errors with 0 at the end, now display it as (!0!) at the end. (Avoid Warnings)

8) Use %zu+(size_t) directly instead of %lu+(uint64_t) convertions.

9) Make sure no warnings are shown in GCC (CodeLite compilation -O3;-std=c++11;-Wall)

10) Updated ReadMe File, with languages colors.

11) Updating Python thus to use more understandable directories. Check examples
to see the difference! Minor corrections on C++ too.

+++ C++/test/t00_Basics.cpp
@@ -87,12 +87,12 @@ printf("\n\n**** TEST:: Error must occur at Line [%i]! *******",__LINE__+1);
 	
 	
 	size_t R0 = CCEXP::Rows(DBG, "T_U8");
-	printf("\nTable T_U8 has [%lu] rows\n", (uint64_t)R0);
+	printf("\nTable T_U8 has [%zu] rows\n", R0);
 	
 		
 	size_t SelRow = 255;
 	size_t C0 = CCEXP::Cols(DBG, "T_U8", SelRow);
-	printf("\nRow [%lu] of Table T_U8 has [%lu] columns\n", SelRow, (size_t)C0);
+	printf("\nRow [%zu] of Table T_U8 has [%zu] columns\n", SelRow, (size_t)C0);
 	
 	// After TestMaxR table, no more errors should exist in the following test.
 	_DBG_ERROR_STOP_OR_CONTINUE_(DBG);
@@ -186,11 +186,11 @@ printf("\n\n**** TEST:: Error must occur at Line [%i]! *******",__LINE__+1);
 	printf("\n\n");
 	char* TableName = NULL;
 	TableName = CCEXP::getTableName(DBG,3);
-	printf("Table Name of Table with ID (%lu) is: %s\n", (uint64_t)3, TableName);
+	printf("Table Name of Table with ID (%i) is: %s\n", 3, TableName);
 	TableName = CCEXP::getTableName(DBG,5);
-	printf("Table Name of Table with ID (%lu) is: %s\n", (uint64_t)5, TableName);
+	printf("Table Name of Table with ID (%i) is: %s\n", 5, TableName);
 	TableName = CCEXP::getTableName(DBG,8);
-	printf("Table Name of Table with ID (%lu) is: %s\n", (uint64_t)8, TableName);
+	printf("Table Name of Table with ID (%i) is: %s\n", 8, TableName);
